Adding workflow for automated creation of draft releases.
This workflow is run when a tag matching the regular expression v*.*.*
is pushed to the GitHub repository. It runs the create_release_notes
script (see ibex_imaging_knowledge_base_utilities package/repository)
and then creates a draft release with the notes as an initial starting
point for the complete release notes.